partitions funcs _topics -> _partitions
