+++ kafka/admin/_partitions.py
@@ -1,4 +1,7 @@
-"""Partition management mixin for KafkaAdminClient."""
+"""Partition management mixin for KafkaAdminClient.
+
+Also defines NewPartitions data class.
+"""
 
 from __future__ import annotations
 
@@ -21,6 +24,60 @@ class PartitionAdminMixin:
     """Mixin providing partition and record management methods."""
     _manager: KafkaConnectionManager
     config: dict
+
+    @staticmethod
+    def _process_create_partitions_input(topic_partitions):
+        _Topic = CreatePartitionsRequest.CreatePartitionsTopic
+        _Assignment = CreatePartitionsRequest.CreatePartitionsTopic.CreatePartitionsAssignment
+        topics = []
+        for topic, count in topic_partitions.items():
+            if isinstance(count, int):
+                topics.append(_Topic(name=topic, count=count))
+            elif isinstance(count, dict):
+                topics.append(
+                    _Topic(
+                        name=topic,
+                        count=count['count'],
+                        assignments=[_Assignment(broker_ids=broker_ids)
+                                     for broker_ids in count['assignments']]))
+            else:
+                topics.append(
+                    _Topic(
+                        name=topic,
+                        count=count.total_count,
+                        assignments=[_Assignment(broker_ids=broker_ids)
+                                     for broker_ids in count.new_assignments]))
+        return topics
+
+    def create_partitions(self, topic_partitions, timeout_ms=None, validate_only=False, raise_errors=True):
+        """Create additional partitions for an existing topic.
+
+        Arguments:
+            topic_partitions: A dict of topic name strings to total partition count (int),
+                or a dict of {topic_name: {count: int, assignments: [[broker_ids]]}}
+                if manual assignment is desired.
+                dict of {topic_name: NewPartitions} is deprecated.
+
+        Keyword Arguments:
+            timeout_ms (numeric, optional): Milliseconds to wait for new partitions to be
+                created before the broker returns.
+            validate_only (bool, optional): If True, don't actually create new partitions.
+                Default: False
+            raise_errors (bool, optional): Whether to raise errors as exceptions. Default True.
+
+        Returns:
+            Appropriate version of CreatePartitionsResponse class.
+        """
+        timeout_ms = self._validate_timeout(timeout_ms)
+        request = CreatePartitionsRequest(
+            topics=self._process_create_partitions_input(topic_partitions),
+            timeout_ms=timeout_ms,
+            validate_only=validate_only)
+
+        def response_errors(r):
+            for result in r.results:
+                yield Errors.for_code(result.error_code)
+        return self._manager.run(self._send_request_to_controller, request, response_errors, raise_errors)
 
     async def _async_get_leader_for_partitions(self, partitions):
         """Finds ID of the leader node for every given topic partition."""
@@ -163,3 +220,21 @@ class PartitionAdminMixin:
         ignore_errors = (Errors.ElectionNotNeededError,)
         return self._manager.run(self._send_request_to_controller, request, response_errors, raise_errors, ignore_errors)
 
+
+class NewPartitions:
+    """DEPRECATED: A class for new partition creation on existing topics.
+
+    Note that the length of new_assignments, if specified, must be the
+    difference between the new total number of partitions and the existing
+    number of partitions.
+
+    Arguments:
+        total_count (int): the total number of partitions that should exist
+            on the topic
+        new_assignments ([[int]]): an array of arrays of replica assignments
+            for new partitions. If not set, broker assigns replicas per an
+            internal algorithm.
+    """
+    def __init__(self, total_count, new_assignments=None):
+        self.total_count = total_count
+        self.new_assignments = new_assignments
